renderer: move GlyphCache into separate file
